// frontend/src/app/shared/api/model/get-order200-response.ts
/**
 * Food Driver API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderStatus } from './order-status';
import { GetOrder200ResponseRestaurant } from './get-order200-response-restaurant';
import { GetOrder200ResponseClient } from './get-order200-response-client';
import { GetOrder200ResponseDeliverer } from './get-order200-response-deliverer';
import { GetOrder200ResponseOrderItemsInner } from './get-order200-response-order-items-inner';

export interface GetOrder200Response {
  [key: string]: any | any;

  /**
   * Unikalny identyfikator
   */
  idOrder: number;
  status: OrderStatus;
  street: string;
  building: string;
  apartment: string;
  postalCode: string;
  city: string;
  payment: string;
  reported: boolean;
  created: string;
  restaurant: GetOrder200ResponseRestaurant;
  client: GetOrder200ResponseClient;
  deliverer: GetOrder200ResponseDeliverer | null;
  orderItems: Array<GetOrder200ResponseOrderItemsInner>;
}
export namespace GetOrder200Response {}