// frontend/src/app/shared/api/model/order-status.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type OrderStatus = 'Nowe' | 'Przyjęte do realizacji' | 'Gotowe do dostawy' | 'Odebrane do dostarczenia' | 'Dostarczone' | 'Zrealizowane' | 'Anulowane';

export const OrderStatus = {
    Nowe: 'Nowe' as OrderStatus,
    PrzyjteDoRealizacji: 'Przyjęte do realizacji' as OrderStatus,
    GotoweDoDostawy: 'Gotowe do dostawy' as OrderStatus,
    OdebraneDoDostarczenia: 'Odebrane do dostarczenia' as OrderStatus,
    Dostarczone: 'Dostarczone' as OrderStatus,
    Zrealizowane: 'Zrealizowane' as OrderStatus,
    Anulowane: 'Anulowane' as OrderStatus
};